#14095: Move ccl unit tests to separate directory
