+++ tests/nightly/t3000/ccl/test_all_gather_nightly.py
@@ -1,1 +1,1 @@
-../../../ttnn/unit_tests/operations/test_all_gather_nightly.py
+../../../ttnn/unit_tests/operations/ccl/test_all_gather_nightly.py
