+++ tests/nightly/t3000/ccl/test_reduce_scatter_nightly.py
@@ -1,1 +1,1 @@
-../../../ttnn/unit_tests/operations/test_reduce_scatter_nightly.py
+../../../ttnn/unit_tests/operations/ccl/test_reduce_scatter_nightly.py
