+++ tests/scripts/tg/run_tg_frequent_tests.sh
@@ -12,7 +12,7 @@ run_tg_tests() {
   pytest -n auto models/demos/tg/llama3_70b/tests/test_llama_decoder_galaxy.py --timeout=600 ; fail+=$?
   pytest -n auto models/demos/tg/llama3_70b/tests/test_llama_model_galaxy_ci.py --timeout=800 ; fail+=$?
   pytest -n auto models/demos/tg/resnet50/tests/test_resnet50_performant.py ; fail+=$?
-  pytest -n auto tests/ttnn/unit_tests/operations/test_all_gather_TG_post_commit.py --timeout=300 ; fail+=$?
+  pytest -n auto tests/ttnn/unit_tests/operations/ccl/test_all_gather_TG_post_commit.py --timeout=300 ; fail+=$?
 
   if [[ $fail -ne 0 ]]; then
     echo "LOG_METAL: run_tg_frequent_tests failed"
