+++ tests/sweep_framework/sweeps/ccl/line_all_gather.py
@@ -11,7 +11,7 @@ import ttnn
 from tests.ttnn.utils_for_testing import start_measuring_time, stop_measuring_time
 from loguru import logger
 from tests.tt_eager.python_api_testing.sweep_tests.comparison_funcs import comp_equal, comp_pcc
-from tests.ttnn.unit_tests.operations.test_all_gather import is_unsupported_case
+from tests.ttnn.unit_tests.operations.ccl.test_all_gather import is_unsupported_case
 from ttnn import ShardTensorToMesh
 
 # Override the default timeout in seconds for hang detection.
