+++ tests/ttnn/unit_tests/operations/ccl/test_all_gather_N300_post_commit.py
@@ -5,7 +5,7 @@
 import pytest
 import ttnn
 from models.utility_functions import skip_for_grayskull
-from tests.ttnn.unit_tests.operations.test_all_gather import (
+from tests.ttnn.unit_tests.operations.ccl.test_all_gather import (
     run_all_gather_on_n300_impl,
     run_all_gather_sharded_n300,
 )
