+++ tests/ttnn/unit_tests/operations/ccl/test_all_gather_matmul.py
@@ -9,7 +9,7 @@ import ttnn
 from ttnn import ShardTensorToMesh, ConcatMeshToTensor
 from tests.tt_eager.python_api_testing.sweep_tests.comparison_funcs import comp_equal, comp_pcc
 from models.utility_functions import skip_for_grayskull, skip_for_wormhole_b0
-from tests.ttnn.unit_tests.operations.test_all_gather import is_unsupported_case
+from tests.ttnn.unit_tests.operations.ccl.test_all_gather import is_unsupported_case
 
 
 USE_NON_FUSED = False
