+++ tests/ttnn/unit_tests/operations/ccl/test_all_gather_nightly.py
@@ -8,7 +8,7 @@ from loguru import logger
 import ttnn
 from tests.tt_eager.python_api_testing.sweep_tests.comparison_funcs import comp_equal, comp_pcc
 from models.utility_functions import skip_for_grayskull
-from tests.ttnn.unit_tests.operations.test_all_gather import (
+from tests.ttnn.unit_tests.operations.ccl.test_all_gather import (
     is_unsupported_case,
     run_all_gather_on_t3000_impl,
 )
