+++ tests/ttnn/unit_tests/operations/ccl/test_reduce_scatter_N300_post_commit.py
@@ -5,7 +5,7 @@
 import pytest
 import ttnn
 from models.utility_functions import skip_for_grayskull
-from tests.ttnn.unit_tests.operations.test_reduce_scatter_post_commit import (
+from tests.ttnn.unit_tests.operations.ccl.test_reduce_scatter_post_commit import (
     run_reduce_scatter_test,
     run_reduce_scatter_sharded_test,
 )
