+++ tests/ttnn/unit_tests/operations/ccl/test_reduce_scatter_llama_perf_sweep.py
@@ -8,7 +8,7 @@ from loguru import logger
 import ttnn
 from tests.tt_eager.python_api_testing.sweep_tests.comparison_funcs import comp_equal, comp_pcc
 from models.utility_functions import skip_for_grayskull, get_devices_for_t3000
-from tests.ttnn.unit_tests.operations.test_reduce_scatter_post_commit import run_reduce_scatter_sharded_test
+from tests.ttnn.unit_tests.operations.ccl.test_reduce_scatter_post_commit import run_reduce_scatter_sharded_test
 
 
 @pytest.mark.timeout(120)
