+++ tests/ttnn/unit_tests/operations/ccl/test_reduce_scatter_nightly.py
@@ -4,7 +4,7 @@
 
 import pytest
 import ttnn
-from tests.ttnn.unit_tests.operations.test_reduce_scatter_post_commit import (
+from tests.ttnn.unit_tests.operations.ccl.test_reduce_scatter_post_commit import (
     run_reduce_scatter_test,
 )
 from models.utility_functions import skip_for_grayskull
